remove my applications page
