+++ apps/web/src/components/Router.tsx
@@ -195,14 +195,6 @@ const CreateApplicationPage = React.lazy(() =>
       ),
   ),
 );
-const MyApplicationsPage = React.lazy(() =>
-  lazyRetry(
-    () =>
-      import(
-        /* webpackChunkName: "tsMyApplicationsPage" */ "../pages/Applications/MyApplicationsPage/MyApplicationsPage"
-      ),
-  ),
-);
 const ApplicationLayout = React.lazy(() =>
   lazyRetry(
     () =>
@@ -836,17 +828,6 @@ const createRoute = (locale: Locales, loginPath: string) =>
                           ],
                         },
                       ],
-                    },
-                    {
-                      path: "applications",
-                      element: (
-                        <RequireAuth
-                          roles={[ROLE_NAME.Applicant]}
-                          loginPath={loginPath}
-                        >
-                          <MyApplicationsPage />
-                        </RequireAuth>
-                      ),
                     },
                   ],
                 },
