--- apps/web/src/pages/Applications/MyApplicationsPage/applicationOperations.graphql
@@ -1,49 +1,0 @@
-query MyApplications {
-  me {
-    id
-    poolCandidates {
-      id
-      status
-      archivedAt
-      submittedAt
-      pool {
-        id
-        closingDate
-        name {
-          en
-          fr
-        }
-        stream
-        classifications {
-          id
-          group
-          level
-          name {
-            en
-            fr
-          }
-          genericJobTitles {
-            id
-            key
-            name {
-              en
-              fr
-            }
-          }
-          minSalary
-          maxSalary
-        }
-      }
-    }
-  }
-}
-
-mutation DeleteApplication($id: ID!) {
-  deleteApplication(id: $id)
-}
-
-mutation ArchiveApplication($id: ID!) {
-  archiveApplication(id: $id) {
-    id
-  }
-}
