+++ apps/web/src/utils/poolCandidate.ts
@@ -1,4 +1,11 @@
 import { IntlShape } from "react-intl";
+import isPast from "date-fns/isPast";
+
+import {
+  formatDate,
+  parseDateTimeUtc,
+  relativeClosingDate,
+} from "@gc-digital-talent/date-helpers";
 
 import { Maybe, PoolCandidateStatus, PublishingGroup } from "~/api/generated";
 import poolCandidateMessages from "~/messages/poolCandidateMessages";
@@ -17,3 +24,42 @@ export const getRecruitmentType = (
   isOngoingPublishingGroup(publishingGroup)
     ? intl.formatMessage(poolCandidateMessages.ongoingRecruitment)
     : intl.formatMessage(poolCandidateMessages.targetedRecruitment);
+
+export const isDraft = (status: Maybe<PoolCandidateStatus>): boolean => {
+  return status === PoolCandidateStatus.Draft;
+};
+
+export const isExpired = (
+  status: Maybe<PoolCandidateStatus>,
+  expirationDate: Maybe<string>,
+): boolean => {
+  if (status === PoolCandidateStatus.Expired) {
+    return true;
+  }
+  return expirationDate ? isPast(parseDateTimeUtc(expirationDate)) : false;
+};
+
+export const formatClosingDate = (
+  closingDate: Maybe<string>,
+  intl: IntlShape,
+): string => {
+  return closingDate
+    ? relativeClosingDate({
+        closingDate: parseDateTimeUtc(closingDate),
+        intl,
+      })
+    : "";
+};
+
+export const formatSubmittedAt = (
+  submittedAt: Maybe<string>,
+  intl: IntlShape,
+): string => {
+  return submittedAt
+    ? formatDate({
+        date: parseDateTimeUtc(submittedAt),
+        formatString: "PPP p",
+        intl,
+      })
+    : "";
+};
