+++ infrastructure/conf/nginx-conf-deploy/default
@@ -195,10 +195,15 @@ server {
         rewrite "^(?:/[a-z]{2})?/directive-on-digital-talent(/.*|$)" /apps/web/dist/index.html break;
     }
 
-    # rewrite possible lang prefix then "applications" to apps/web index.html
-    location ~ "^(?:/[a-z]{2})?/applications(/.*|$)" {
-          add_header Cache-Control "no-cache, max-age=0"; # no-caching - can cause chunk-not-found on rebuild
-        rewrite "^(?:/[a-z]{2})?/applications(/.*|$)" /apps/web/dist/index.html break;
+    # Redirect old "applications" to new profile and applications
+    location = /en/applications {
+      absolute_redirect off;
+      return 301 /en/applicant/profile-and-applications$is_args$args;
+    }
+
+    location = /fr/applications {
+        absolute_redirect off;
+        return 301 /fr/applicant/profile-and-applications$is_args$args;
     }
 
     # Redirect old dashboard to new profile and applications
