+++ infrastructure/conf/nginx-conf-local/default
@@ -213,9 +213,15 @@ server {
         rewrite "^(?:/[a-z]{2})?/directive-on-digital-talent(/.*|$)" /apps/web/dist/index.html break;
     }
 
-    # rewrite possible lang prefix then "applications" to apps/web index.html
-    location ~ "^(?:/[a-z]{2})?/applications(/.*|$)" {
-        rewrite "^(?:/[a-z]{2})?/applications(/.*|$)" /apps/web/dist/index.html break;
+    # Redirect old "applications" to new profile and applications
+    location = /en/applications {
+      absolute_redirect off;
+      return 301 /en/applicant/profile-and-applications$is_args$args;
+    }
+
+    location = /fr/applications {
+        absolute_redirect off;
+        return 301 /fr/applicant/profile-and-applications$is_args$args;
     }
 
     # Redirect old dashboard to new profile and applications
